Move some ui tools into tools
