+++ trview.app/Tools/Compass.cpp
