+++ trview.app/Tools/Compass.h
