+++ trview.app/Tools/Measure.cpp
