+++ trview.app/Tools/Measure.h
