+++ trview.app/Tools/Toolbar.cpp
