+++ trview.app/Tools/Toolbar.h
